+++ src/main/java/services/place/PlaceWebSocket.java
@@ -1,4 +1,4 @@
-package services;
+package services.place;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
